feat: tell the user when calling a tool/function
This is meant to make the AI assistant appear more responsive to the user, given that the tool call and digesting its response may take a moment and that the LLM calls multiple times.
Also, this clarifies to the user whether and which functions the AI uses to answer a request.

Additional changes:
* Tell the user about a chat completion error such as a timeout.
* Increase the chat completion request timeout (because otherwise large search results or wikipedia articles result in the request to timeout on my machine)