refactor: consistently use AudioMessage for in/out
Change the internal interface to use the AudioMessage struct for both input and output in order to pass through the input audio as is.
This is also to open the door for experimentation with multi-modal models such as qwen2.5-omni.
However, turns out that there are still issues preventing multi-modal models from working:
* langchaingo doesn't support audio messages yet, see https://github.com/tmc/langchaingo/issues/958.
* localai doesn't support audio messages properly (uses `audio_url` instead of `input_audio`), see https://github.com/mudler/LocalAI/issues/5226#issuecomment-3075370658 and https://platform.openai.com/docs/api-reference/chat/create.